create voting system

+++ app/router/post.py
@@ -1,5 +1,5 @@
 
-from typing import List
+from typing import List, Optional
 from fastapi import APIRouter, Body, Depends, HTTPException, Response, status
 from sqlalchemy.orm import Session
 
@@ -18,7 +18,8 @@ router = APIRouter(
 
 @router.get("/", response_model=List[Post])
 # Fetch all posts from the database, with an in-memory fallback if the query fails.
-async def get_posts(db: Session = Depends(get_db), get_current_user: int = Depends(oauth2.get_current_user)):
+async def get_posts(db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user),
+                    limit: int = 10, skip: int = 0, search: Optional[str] = ""):
     # try:
     #     conn = psycopg2.connect(host='localhost', database='fastapi', user='postgres', password='password', cursor_factory=RealDictCursor)
     #     cursor = conn.cursor()
@@ -30,7 +31,8 @@ async def get_posts(db: Session = Depends(get_db), get_current_user: int = Depen
     # except Exception as error:
     #     print("Error: ", error)
     #     return {"data": my_posts}
-    posts = db.query(models.Post).all()
+    # posts = db.query(models.Post).filter(models.Post.owner_id == current_user.id).limit(limit).offset(skip).all()
+    posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 
@@ -40,14 +42,14 @@ async def get_posts(db: Session = Depends(get_db), get_current_user: int = Depen
 @router.post("/create_post", status_code=status.HTTP_201_CREATED, response_model=Post)
 # Create a new post in the database and return the inserted record.
 async def create_post(post: PostCreate = Body(...), db: Session = Depends(get_db), 
-                      get_current_user: int = Depends(oauth2.get_current_user)):
+                      current_user: models.User = Depends(oauth2.get_current_user)):
     # cursor.execute("INSERT INTO posts (title, content, published) VALUES (%s, %s, %s) RETURNING *",
     #                 (post.title, post.content, post.published))
     # new_post = cursor.fetchone()
     # conn.commit()
     # print(new_post)
     # new_post = models.Post(title=post.title, content=post.content, published=post.published)
-    new_post = models.Post(**post.model_dump())
+    new_post = models.Post(owner_id=current_user.id, **post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -58,16 +60,19 @@ async def create_post(post: PostCreate = Body(...), db: Session = Depends(get_db
 
 @router.get("/latest")
 # Return the most recent post from the in-memory posts list.
-async def get_latest_post(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+async def get_latest_post(db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     posts = db.query(models.Post).all()
     return  posts[-1] if posts else None
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT, response_class=Response)
 # Delete a post by id and return 404 if the post does not exist.
-async def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+async def delete_post(id: int, db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id: {id} not found")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to delete this post")    
     db.delete(post, synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)                 
@@ -81,7 +86,7 @@ async def delete_post(id: int, db: Session = Depends(get_db), current_user: int 
 
 @router.put("/{id}", response_model=Post)
 # Update a database post by id and return the updated post.
-async def update_post(id: int, updated_post: PostUpdate = Body(...), db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+async def update_post(id: int, updated_post: PostUpdate = Body(...), db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     # cursor.execute("UPDATE posts SET title = %s, content = %s, published = %s WHERE id = %s RETURNING *",
     #                 (post.title, post.content, post.published, str(id)))
     # updated_post = cursor.fetchone()
@@ -99,9 +104,12 @@ async def update_post(id: int, updated_post: PostUpdate = Body(...), db: Session
 
 @router.get("/{id}", response_model=Post)
 # Fetch one post by id from the database and return 404 if it is missing.
-async def get_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+async def get_post(id: int, db: Session = Depends(get_db), current_user: models.User = Depends(oauth2.get_current_user)):
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id: {id} not found")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to view this post")
     return post
